nothing of note

+++ src/test.cpp
@@ -28,7 +28,7 @@ if the last bit is 0
 int main(void){
     int8 hi = -128;
 
-    std::cout << (int)max(hi) << N << min(hi);
+    std::cout << (int)max(hi) << N << (int)min(hi);
     test();
